+++ nyquist/portaudio/src/os/win/pa_win_util.c
@@ -62,7 +62,7 @@ static int numAllocations_ = 0;
 #endif
 
 
-void *PaUtil_AllocateMemory( long size )
+void *PaUtil_AllocateMemory( size_t size )
 {
     void *result = GlobalAlloc( GPTR, size );
